ncfd/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ncfd"
version = "0.1.0"
description = "Near-Certain Failure Detector for US-listed biotech pivotal trials"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = {text = "MIT"}
authors = [{ name = "Your Team" }]

dependencies = [
  "requests>=2.31",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "SQLAlchemy>=2.0",
  "psycopg2-binary>=2.9",
  "duckdb>=1.0",
  "prefect>=2.16",
  "beautifulsoup4>=4.12",
  "lxml>=5.0",
  "PyYAML>=6.0",
  "alembic>=1.13",
  "python-dateutil>=2.9",
  "tqdm>=4.66",
  "uvloop>=0.19; platform_system != 'Windows'",
  "pandas>=2.0",
  "numpy>=1.25",
  "openai>=1.0",
]

[project.optional-dependencies]
api = ["fastapi>=0.112", "uvicorn[standard]>=0.30"]
extract = ["langextract>=0"]  # if private/not on PyPI, pin as VCS in your env
parsers = [
  "beautifulsoup4>=4.12",
  "lxml>=5",
  "pdfminer.six>=20231228"
]

dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "freezegun>=1.5",
  "ruff>=0.5",
  "black>=24.4",
  "mypy>=1.10",
  "types-requests",
  "pre-commit>=3.7",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
select = ["E","F","I","B","UP"]
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = []

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --disable-warnings --maxfail=1"

[tool.coverage.run]
source = ["src/ncfd"]
branch = true

[tool.coverage.report]
skip_empty = true
